fix: converting types to ensure xgboost compatibility

+++ peptide_forest/training.py
@@ -373,8 +373,8 @@ def fit_cv(df, score_col, cv_split_data, sensitivity, q_cut, conf):
         feature_importances.append(rfreg.feature_importances_)
 
         # Score predictions
-        scores_train = rfreg.score_psms(train[features])
-        scores_test = rfreg.score_psms(test[features])
+        scores_train = rfreg.score_psms(train[features].astype(float))
+        scores_test = rfreg.score_psms(test[features].astype(float))
         df.loc[train.index, "prev_score_train"] = scores_train
         df.loc[train.index, "model_score_train"] += scores_train
         df.loc[test.index, "model_score"] = scores_test
